TodoListAPI/requirements.txt
asgiref==3.8.1
Django==5.2.3
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
